python-service/requirements.txt
# ========================================
# Brainstorming Module - Required Packages
# Python 3.10+ required
# ========================================

# ========================================
# Web Framework
# ========================================
fastapi==0.115.0
uvicorn[standard]==0.32.0

# ========================================
# Configuration
# ========================================
python-dotenv==1.0.1
pydantic==2.12.4
pydantic-settings==2.6.1

# ========================================
# OpenAI & LLM
# ========================================
openai==1.57.0

# ========================================
# LangChain & RAG
# ========================================
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.79
langchain-openai==0.2.13
langchain-text-splitters==0.3.11

# ========================================
# Vector Database
# ========================================
chromadb==0.5.23

# ========================================
# Utilities
# ========================================
python-dateutil==2.9.0.post0

# ========================================
# Notes
# ========================================
# 브레인스토밍 모듈 전용 (2025-11-28)
# - chromadb==0.5.23: 팀플 ChromaDB 데이터 호환
# - langchain==0.3.27: RAG 시스템
# - openai==1.57.0: GPT-4o 사용
# - pydantic==2.12.4: 팀플과 동일 버전
#
# 제외된 패키지:
# - Database & ORM: 메모리 세션 관리만 사용
# - Authentication: Java Spring Boot에서 처리
# - PDF Processing: 브레인스토밍에서 미사용
# - Testing: 개발용으로 일단 제외
